requirements.txt
# Core Dependencies
streamlit>=1.28.0
python-docx>=0.8.11
python-multipart

# Document Processing
mammoth>=1.6.0
python-pptx
PyPDF2

# Database Dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
psycopg[pool]>=3.1.0
alembic>=1.12.0
sqlalchemy-utils>=0.41.0

# Distributed task queue
celery>=5.3
redis>=4.0

# System & Performance
psutil
python-magic
python-dateutil
pytz

# Security
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.4

# Data Processing
pandas
numpy

# Concurrency
concurrent-log-handler

# Development Tools
python-dotenv
structlog>=23.0.0
tqdm
colorama

# Email & Communication
yagmail
